Fix coverage
As far as I can tell, the problem is that we're computing coverage
independently in the unit, integration and acceptance tests.

Different solutions include:
- Always putting the --cov-append flag, but it might lead to erroneous
reports
- Switching matrix to individual runs. It could be interesting to
identify erroneous builds quicker but at the price of setupping the VM
more times (and it seems it's actually the longest part of the CI)
- Making it so that --cov-append is only passed in the context of the
CI, which is what we're doing here.