Close race condition in procrastinate_fetch_job
This closes a race condition in the procrastinate_fetch_job plpgsql function,
where jobs sharing the same lock can be run out of order.

With this commit jobs with the same lock are **always** executed in order,
whatever their ETAs and queues.

In effect:
- if job A in queue 1 (id 1) and job B in queue 2 (id 2) have the same lock,
  and no workers process queue 1, then job B won't be executed, because
  job A must be executed first
- if job A is deferred with ETA 1 year, no other jobs with the same lock will
  be executed for 1 year

The lock name may change from "lock" to "serial lock" in the future.