Fix the procrastinate_defer_periodic_job SQL function
Concurent DELETE's may result in deadlocks because the concurrent transactions
may not acquire the share locks on the rows to delete in a consistent order.

This problem can be worked out by using USING with a SELECT ORDER BY FOR UPDATE
guaranteeing that share locks are acquired in a consistent order.